discord-music-bot/requirements.txt
discord.py
youtube-dl
pydub
ffmpeg-python
spotipy
soundcloud-python
apple-music-python
aiohttp